Implemented patient features and changed login

+++ HealthInstitution/GUI/PatientView/PatientScheduleWindow.xaml.cs
@@ -1,20 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-using HealthInstitution.Core.SystemUsers.Users.Model;
-using HealthInstitution.Core.Examinations.Model;
+﻿using HealthInstitution.Core.Examinations.Model;
 using HealthInstitution.Core.Examinations.Repository;
+using HealthInstitution.Core.ScheduleEditRequests.Repository;
+using HealthInstitution.Core.SystemUsers.Users.Model;
+using HealthInstitution.Core.TrollCounters.Repository;
 using HealthInstitution.GUI.PatientView;
+using System.Windows;
 
 namespace HealthInstitution.GUI.PatientWindows;
 
@@ -34,29 +24,81 @@ public partial class PatientScheduleWindow : Window
 
     private void addButton_click(object sender, RoutedEventArgs e)
     {
-        AddExaminationDialog addExaminationDialog = new AddExaminationDialog();
-        addExaminationDialog.ShowDialog();
+        try
+        {
+            TrollCounterFileRepository.GetInstance().TrollCheck(loggedPatient.username);
+            AddExaminationDialog addExaminationDialog = new AddExaminationDialog(loggedPatient);
+            addExaminationDialog.ShowDialog();
+            dataGrid.Items.Clear();
+            LoadGrid();
+            TrollCounterFileRepository.GetInstance().GetTrollCounterById(loggedPatient.username).AppendCreateDates(DateTime.Today);
+            TrollCounterFileRepository.GetInstance().Save();
+        }
+        catch (Exception ex)
+        {
+            System.Windows.MessageBox.Show(ex.Message, "Troll Alert", MessageBoxButton.OK, MessageBoxImage.Error);
+            this.Close();
+        }
     }
 
     private void editButton_click(object sender, RoutedEventArgs e)
     {
-        Examination selectedExamination = (Examination)dataGrid.SelectedItem;
-        EditExaminationDialog editExaminationDialog = new EditExaminationDialog(selectedExamination);
-        editExaminationDialog.ShowDialog();
+        try
+        {
+            TrollCounterFileRepository.GetInstance().TrollCheck(loggedPatient.username);
+            Examination selectedExamination = (Examination)dataGrid.SelectedItem;
+            EditExaminationDialog editExaminationDialog = new EditExaminationDialog(selectedExamination);
+            editExaminationDialog.ShowDialog();
+            dataGrid.Items.Clear();
+            LoadGrid();
+            TrollCounterFileRepository.GetInstance().GetTrollCounterById(loggedPatient.username).AppendEditDeleteDates(DateTime.Today);
+            TrollCounterFileRepository.GetInstance().Save();
+        }
+        catch (Exception ex)
+        {
+            System.Windows.MessageBox.Show(ex.Message, "Troll Alert", MessageBoxButton.OK, MessageBoxImage.Error);
+            this.Close();
+        }
     }
 
     private void deleteButton_click(object sender, RoutedEventArgs e)
     {
+        Examination selectedExamination = (Examination)dataGrid.SelectedItem;
+        try
+        {
+            TrollCounterFileRepository.GetInstance().TrollCheck(loggedPatient.username);
+            ExaminationDoctorRepository.GetInstance().SaveExaminationDoctor();
+            dataGrid.Items.Clear();
+            LoadGrid();
+            TrollCounterFileRepository.GetInstance().GetTrollCounterById(loggedPatient.username).AppendEditDeleteDates(DateTime.Today);
+            TrollCounterFileRepository.GetInstance().Save();
+        }
+        catch (Exception ex)
+        {
+            System.Windows.MessageBox.Show(ex.Message, "Troll Alert", MessageBoxButton.OK, MessageBoxImage.Error);
+            this.Close();
+        }
         if (System.Windows.MessageBox.Show("Are you sure you want to delete selected examination", "Question", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
         {
-            Examination selectedExamination = (Examination)dataGrid.SelectedItem;
-            dataGrid.Items.Remove(selectedExamination);
-            examinationRepository.DeleteExamination(selectedExamination.id);
-            //dodaj isto za doktora.
+            if (selectedExamination.appointment.AddDays(-2) < DateTime.Now)
+            {
+                ScheduleEditRequestFileRepository.GetInstance().AddDeleteRequest(selectedExamination);
+            }
+            else
+            {
+                dataGrid.Items.Remove(selectedExamination);
+                examinationRepository.DeleteExamination(selectedExamination.id);
+                selectedExamination.doctor.examinations.Remove(selectedExamination);
+            }
         }
     }
 
     private void dataGrid_Loaded(object sender, RoutedEventArgs e)
+    {
+        LoadGrid();
+    }
+
+    private void LoadGrid()
     {
         foreach (Examination examination in ExaminationRepository.GetInstance().examinations)
         {
